requirements.txt
streamlit
langchain
openai
pydantic
python-dotenv
